chore: split catalog internals into smaller files
